Work around read() not finishing under paste
When run from mod_wsgi, env["wsgi.input"].read's default behavior of
reading until the end of the data is reached pulls in the rest of the
client's request.  Under paste with SSL, it blocks.  So if the client
sent a Content-Length header, stop after reading just that many bytes.