Downgrade socket problems to warnings
Previously, these were logged at exception - which logs at ERROR and
prints a traceback.  This led to two problems: first, that they're not
kdcproxy errors (rather problems with the other end); and second, that
the traceback is quite noisy.  Log a simplified version of the
exception instead of the traceback.

In the process, correct the sendall() error message to refer to
sendall(), not recv().